Rename terminalImportMap into manualImportMap, remove initialImportMap
