+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@jsenv/importmap-node-module",
-  "version": "2.5.0",
+  "version": "2.6.0",
   "description": "Generate importmap for node_modules",
   "license": "MIT",
   "repository": {
