+++ test/mapping_overrides/mapping_overrides.test.js
@@ -26,42 +26,10 @@ const test = async (options) => {
   return importmaps["test.importmap"]
 }
 
-// initialImportMap is not enough (it gets overridden by package.json)
+// manualImportMap allows to override the mapping found in package.json
 {
   const actual = await test({
-    initialImportMap: {
-      scopes: {
-        "./node_modules/react-redux/": {
-          react: "./node_modules/preact/compat/src/index.js",
-        },
-      },
-    },
-  })
-  const expected = {
-    imports: {
-      "react-redux/": "./node_modules/react-redux/",
-      "react-redux": "./node_modules/react-redux/index.js",
-      "root/": "./",
-      "root": "./index.js",
-    },
-    scopes: {
-      "./node_modules/react-redux/": {
-        // Here we would like to find "preact" instead of "react"
-        "react/": "./node_modules/react/",
-        "react": "./node_modules/react/index.js",
-      },
-      "./node_modules/react/": {
-        "react/": "./node_modules/react/",
-      },
-    },
-  }
-  assert({ actual, expected })
-}
-
-// terminalImportMap allows to override the mapping found in package.json
-{
-  const actual = await test({
-    terminalImportMap: {
+    manualImportMap: {
       scopes: {
         "./node_modules/react-redux/": {
           react: "./node_modules/preact/compat/src/index.js",
