+++ test/updating_jsconfig/updating_jsconfig_react.test.mjs
@@ -28,7 +28,7 @@ await writeImportMapFiles({
   projectDirectoryUrl: testDirectoryUrl,
   importMapFiles: {
     "test.importmap": {
-      initialImportMap: {
+      manualImportMap: {
         imports: { foo: "./bar.js" },
       },
       useForJsConfigJSON: true,
